fixes for compiling under groovy 2.2+

+++ src/main/groovy/Stats.groovy
@@ -235,7 +235,7 @@ class Stats extends DescriptiveStatistics {
     @CompileStatic
     static SummaryStatistics summary(double [] values) {
         int i=0;
-        summary { values[i++] }
+        summary { values[(int)(i++)] }
     }
     
     /**
